Rewrite of the entire simpl Java application in JS (RAW!)
This implementation was done in a few hours, there be dragons...

Hopefully I will soon clean up the codebase and add some JS tests
to verify that the application is working as intended.

This is a rewrite of the simpl application in JS to make it more portable,
I believe that having a simple html page with the complete source code
is much easier to distribute, with the added bonus that the web user interface
is very simple and easy to use.